route/requirements.txt
mysql-connector-python 
pika